pyproject.toml
[tool.poetry]
name = "fish-diffusion"
version = "0.1.0"
description = ""
authors = ["Lengyue <lengyue@lengyue.me>"]
license = "Apache"

packages = [{ include = "fish_diffusion" }]

[tool.poetry.dependencies]
python = "^3.10"
praat-parselmouth = "^0.4.3"
soundfile = "^0.11.0"
librosa = "^0.9.1"
pytorch-lightning = "^1.8.6"
numba = "^0.56.4"
fish-audio-preprocess = "^0.1.9"
wandb = "^0.13.9"
transformers = "^4.25.1"
torchcrepe = "^0.0.17"
mmengine = "^0.4.0"
loguru = "^0.6.0"
click = "^8.1.3"
tensorboard = "^2.11.2"
openai-whisper = "^20230124"
pypinyin = "^0.48.0"
TextGrid = "^1.5"
pyworld = "^0.3.2"
pykakasi = "^2.2.1"
gradio = "^3.18.0"
onnxruntime = "^1.14.0"
flask = "^2.2.3"
flask-cors = "^3.0.10"
fairseq = "^0.12.2"
audiomentations = "^0.28.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.11.4"
black = "^22.12.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
furo = "^2022.12.7"
sphinx-autobuild = "^2021.3.14"
myst-parser = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
extend_skip = ["dataset", "logs"]
